Adds a simple bootstrap script for a local concourse instance (#1)
TL;DR
-----

Implements a simple local install of Concourse for bootstrapping

Details
-------

This change adds the `bootstrap` script that starts a local 
Concourse using Docker Compose. Includes a sample pipeline
and loads that pipeline into that Concourse instance and 
opens a browser to it. The sample pipeline will evolve into the
actual bootstrap pipeline.